requirements.txt
numpy
numba
pygame
scipy
PySDL2
pysdl2-dll